Database---Prediction-Pipeline/requirements.txt
<<<<<<< HEAD
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.4
scikit-learn==1.3.2
python-dotenv==1.0.0
pydantic==2.5.0
joblib==1.3.2 
=======
# Database connectors
mysql-connector-python>=8.0.0
pymongo>=4.0.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# Visualization
matplotlib>=3.5.0

# Logging and utilities
python-dotenv>=0.19.0 
>>>>>>> 2f891f6c837c022373fcdd91d92764e638e9f13e
